After installing sndfile:  - fix SNDFILE_xxx
Give a better failure description for some errors (sizeof check).

If there are intXX_t but no uintXX_t, use the intXX_t with "unsigned" instead
of trying to find the right K&R-C integral type.

Use macho as the format type for NASM on OSX, but disable the nasm use by
default there, as there seems to be a problem with the nasm code on OSX.